Added gscan to renovate ignore list
no ref

The gscan library is one that we intentionally want to keep out of sync whenever we're releasing new theme features, as the latest version is publicly available at themes.ghost.org for theme devs to check their builds. After we give an appropriate lead time for updates, *then* we want to update Ghost to require them.